requirements.txt
Flask==2.2.5
requests==2.31.0
Flask-Cors==3.0.10
gunicorn==20.1.0
python-dotenv==1.0.0
pdf2image==1.17.0
Pillow==10.2.0
